samples: nrf9160: agps: Add PGPS support
Using the prj_pgps.conf Kconfig overlay, enable
PGPS features. This includes MCUBOOT, which is
only needed to define the symbol and address of
PM_MCUBOOT_SECONDARY_ADDRESS.

Add conditionally-compiled hooks to main.c
so that PGPS is initialized, data received from
cloud is processed, a relevant prediction can be
selected, and that prediction injected to the
modem instead of getting it on demand from the
cloud via AGPS.

Make this work with AGPS alone, PGPS alone,
or both together.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>